Add files via upload
The interest and demand in drone are higher than ever. With this popular demand, 
new types of drone merchandise have been designed and manufactured, so that civilians can 
afford to buy them for various purposes. However, as it got easier for drone to be used by 
more people, safety and security issues have raised as accidents are much more likely to 
happen. For safety purposes, it is essential for observers and drone to be aware of an 
approaching drone. In this work, we introduce a comprehensive drone detection and 
classification system based on deep learning. Drone detection is performed using YOLOV2 
and drone classification is performed using convolutional neural networks (CNN). Classified 
types of drones are Tricopter and quadcopter